create and delete images with base64

+++ app/Controllers/ProductController.php
@@ -3,8 +3,10 @@
 namespace App\Controllers;
 
 use App\Models\ProductModel;
+use App\Models\ImageModel;
 use Mindk\Framework\Http\Request\Request;
 use Mindk\Framework\Exceptions\NotFoundException;
+use Mindk\Framework\Auth\AuthService;
 
 /**
  * Product controller
@@ -46,7 +48,7 @@ class ProductController
     /**
      * Create product
      */
-    function create(Request $request, ProductModel $model){
+    function create(Request $request, ProductModel $model, ImageModel $imageModel){
 
         //@TODO: Implement this
         // Accept the assumption that from the form in POST request 
@@ -56,13 +58,26 @@ class ProductController
         // just call: $request->get('storage', null, array)
 
         // Take information, what sendeed in form
+        $user = AuthService::getUser();
         $data = $request->get('formData', null, 'array');
+        $data += ['user_id' => $user->id];
+        $imageData = $request->get('image', null, '');
 
         if(empty($data)) {
             throw new \Exception('No POST data to create item');
         }
 
+        // creating product in DB
         $item = $model->create($data);
+
+        //if creating was successful
+        if ($item->id) {
+            // uploading image
+            if(!empty($imageData["filename"]) and !empty($imageData["body"])) {
+                $imageData += ['product_id' => $item->id];
+                $imageModel->upload($imageData);
+            }
+        }
 
         return $item;
     }
@@ -86,10 +101,18 @@ class ProductController
     /**
      * Delete the product
      */
-    function delete(ProductModel $model, $id) {
-        
-        $item = $model->delete($id);
-        return $item;
+    function delete(ProductModel $model, $id, ImageModel $imageModel) {
 
+        $user = AuthService::getUser();
+        $item = $model->load($id);
+        if ( ($item->user_id == $user->id) or ($user->role == 'admin') ) {
+            if ($imageModel->remove($id)) {
+                $item = $model->delete($id);
+            } else {
+                throw new \Exception('Deleting error - product not deleted');
+            }
+        }
+
+        return $item;
     }
 }
